Add basic calculations and operate function
Some issues came up with the operate function. It's either gonna have to be quite verbose, so the operators become something like "opPlus". or the current implementation has to be used. Now we can use +/- etc, but they have to be strings. This doesn't seem ideal but will come back to it later when the workings of the whole project have become more clear.