Listen to number and AC clicks and update display
Added class to number buttons to make it easier to select them all. These buttons are then selected to listen to clicks. On each click the display is updated. If the display is in standby mode the display is first cleared.

This method might have to change when doing actual calculations and having to save values entered and outputted on the go. So will probably come back to that later