Add class to screen when displaying error message
The class makes sure the screen font-size is changed to be slightly smaller. This is in response to the font for the display by default becoming slightly bigger for readability. This way error messages can still be displayed in full.

Besides this, been playing around with a way to round the numbers that doesn't make me feel bad. toFixed shows up everywhere and seems to be the easy solution, but a purely mathematical way without converting to strings would have been preferred. There is some sample code for a function that mostly achieves this goal, but seems to have some typical floating point issues. For some numbers with round at the last integer, we get some trailing 000001 or similar. From Googling around, this seems to be a common issue with floating point math in js.

All in all I spent way too much time on trying to find a solution that worked with just maths. A good amount of time spent on learning, but also full of frustration.

The new function using toExponential and toFixed is also in the file now. This will be tested further and implemented in later commits.