Add rounding function and call it in evaluate
Imlemented a mathematical way of rounding using the techniques previously explored. Still not ideal I think, and the floating point issue is mildly frustrating to say the least. I feel like there should be some ways to work around this, but that might be a topic to visit later.

Implementing the sign of the value was an interesting unexpected issue popping up, but ended up being relatively easy to solve.

Overall, contemplating the project, there might be some room for optimisation or interesting functionality. Using the input in an array and operate on it using reduce could be a fun exercise for later.

Furthermore a more Android Calculator style of filling the screen with the inputs could be challenging but fun and would probably work best with array reduce. Instead of this, a small second output storing and showing the current operation could be fun as well.

For now, the main thing to do is make sure the sqrt works properly. Besides that, all main and bonus goals seem achieved.